refactor(engine): Break dsl.common into smaller independent units + put dslinput converters in DSLInput
